requirements.txt
python-dotenv==1.0.1
colorama==0.4.6
pytest==7.4.4
pytest-cov==4.1.0
psycopg==3.1.18  # Substituído psycopg2-binary por psycopg v3
